Increase Playwright navigation timeout to 60s for proxy scenarios
Increased all page.Goto() timeouts from 30s to 60s to handle slow proxy connections better.
This should significantly reduce timeout failures in production.

Changes:
- job.go: Added 60s timeout to initial navigation
- place.go: Added 60s timeout to place detail navigation
- hybridjob.go: Increased hybrid mode navigation from 30s to 60s
- nearbysearch.go: Increased nearby search navigation from 30s to 60s
- emailjob.go: Increased email extraction navigation from 30s to 60s

+++ gmaps/emailjob.go
@@ -114,7 +114,7 @@ func (j *EmailExtractJob) BrowserActions(ctx context.Context, page playwright.Pa
 
 	pageResponse, err := page.Goto(j.GetURL(), playwright.PageGotoOptions{
 		WaitUntil: playwright.WaitUntilStateDomcontentloaded,
-		Timeout:   playwright.Float(30000), // 30 second timeout for website loading (increased from 20s)
+		Timeout:   playwright.Float(60000), // 60 second timeout for website loading with proxies
 	})
 	if err != nil {
 		// Distinguish between permanent and temporary failures
